refactor(): User DTO's package refactor
